+++ javascript/react-hooks/src/index.ts
@@ -2,4 +2,9 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './hooks.js'
+export * from './useColumnValueOptions.js'
+export * from './useSimpleDropdownOptions.js'
+export * from './useEnumDropdownOptions.js'
+export * from './useDateFormatPatternOptions.js'
+export * from './useTableNames.js'
+export * from './useTableColumnNames.js'
