+++ javascript/react-hooks/src/useTableColumnNames.ts
@@ -1,0 +1,13 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+import type ColumnTable from 'arquero/dist/types/table/column-table'
+import { useMemo } from 'react'
+i
+export function useTableColumnNames(
+	table: ColumnTable | undefined,
+	filter?: (name: string) => boolean,
+): string[] {
+	return useMemo(() => table?.columnNames(filter) || [], [table, filter])
+}
